fix: refactor the mobile-web mcp server structure
