+++ packages/mcp-provider-mobile-web/src/tools/offline-analysis/ruleConfig.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { CodeAnalysisBaseIssueType } from '../../../schemas/analysisSchema.js';
+import { CodeAnalysisBaseIssueType } from '../../schemas/analysisSchema.js';
 import dedent from 'dedent';
 
 export interface RuleConfig {
