+++ packages/mcp-provider-mobile-web/test/provider.test.ts
@@ -1,8 +1,8 @@
 import { McpProvider, McpTool, Services } from "@salesforce/mcp-provider-api";
 import { MobileWebMcpProvider } from "../src/provider.js";
 import { NativeCapabilityTool } from "../src/tools/native-capabilities/nativeCapabilityTool.js";
-import { OfflineAnalysisTool } from "../src/tools/mobile-offline/offline-analysis/sf-mobile-web-offline-analysis.js";
-import { OfflineGuidanceTool } from "../src/tools/mobile-offline/offline-guidance/sf-mobile-web-offline-guidance.js";
+import { OfflineAnalysisTool } from "../src/tools/offline-analysis/sf-mobile-web-offline-analysis.js";
+import { OfflineGuidanceTool } from "../src/tools/offline-guidance/sf-mobile-web-offline-guidance.js";
 import { StubServices } from "./test-doubles.js";
 
 describe("Tests for MobileWebMcpProvider", () => {
