+++ packages/mcp/src/registry.ts
@@ -17,12 +17,14 @@
 import { McpProvider } from '@salesforce/mcp-provider-api';
 import { DxCoreMcpProvider } from '@salesforce/mcp-provider-dx-core';
 import { CodeAnalyzerMcpProvider } from '@salesforce/mcp-provider-code-analyzer';
+import { MobileWebMcpProvider } from '@salesforce/mcp-provider-mobile-web';
 
 
 /** -------- ADD McpProvider INSTANCES HERE ------------------------------------------------------------------------- */
 
 export const MCP_PROVIDER_REGISTRY: McpProvider[] = [
   new DxCoreMcpProvider(),
-  new CodeAnalyzerMcpProvider()
+  new CodeAnalyzerMcpProvider(),
+  new MobileWebMcpProvider(),
   // Add new instances here
 ];
